requirements.txt
asgiref==3.8.1
Django==4.2.20
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
numpy==2.0.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
